feat: move context window into engine select
