+++ packages/playground/src/components/room/index.ts
@@ -1,5 +1,4 @@
 export * from "./room-content";
-export * from "./room-context-chart";
 export * from "./room-inline-tool";
 export * from "./room-input";
 export * from "./room-input-menu-file-explorer";
